feat: add capability to have multiple trees but a single default tree
The default tree is used as the de facto representative of the perk group and is the one that all the functions use to manipulate or give information about the perk group. The other trees can be gotten manually when needed (primarily for adding alternate trees for this perk group for AI enemies).