feat!: allow category in DynamicMap as function that returns an array
Instead of it being always an array that could contain strings or weighted containers, it must now always be an array of PerkGroup IDs or a function that takes the `_perkTree` as its argument and returns an array of PerkGroup IDs. This allows defining the guaranteed perk groups of the category with more complex logic than before.