fix: during removePerkGroup skip perks belonging to other perk groups
Only remove perks from this perk tree which are added to this perk tree solely from this perk group (i.e. _perkGroupID). If the perk additionally belongs to another perk group which is also present in this perk tree then don't remove it, because that will inadvertently cause that perk group to become incomplete.