fix: don't include nameless perk groups in perk groups of perks
This is similar to how nameless perk groups are excluded in the tooltip of perks. The idea is that only perk groups with names can be used as a reference point for the player in the perk tree. Nameless perk groups can be used for other purposes e.g. a custom perk group for enemies.